물리엔진 적용한 애니메이션(기초)

+++ Take Me Campus/object.h
@@ -9,7 +9,7 @@
 #define FALSE 0
 
 typedef struct {
-	int x, y;		// 오브젝트 좌표
+	float x, y;		// 오브젝트 좌표
 } Position;
 
 typedef struct {
